change the cache timeout value

+++ Portal/src/main/java/org/easystogu/portal/vo/StatisticsVO.java
@@ -2,9 +2,9 @@ package org.easystogu.portal.vo;
 
 public class StatisticsVO {
 	public String date;
-	public int count1;
-	public int count2;
-	public int count3;
-	public int count4;
-	public int count5;
+	public double count1;
+	public double count2;
+	public double count3;
+	public double count4;
+	public double count5;
 }
